Prepare mkmini.sh and friends for Ubuntu 16 (#7)
* Prepare mkmini.sh and friends for Ubuntu 16

Changes:

- Make mkmini.sh agnostic re: Ubuntu releases. It now defaults to
  generating Xenial-based images, but can still be used to make
  Trusty-based ones.
- The generated images are now named scion-codename.iso they don't
  clobber each other.
- Fixed the confirmation prompt during LVM setup against a completely
  empty disk.
- Make the script's on log lines distinguishable by color.

+++ preseed/mkmini.sh
@@ -2,14 +2,15 @@
 
 set -e
 
+FLAVOR=${1:-xenial}
 BUILD_DIR="build"
-MINI_ISO="$BUILD_DIR/mini.iso"
-SCION_ISO="$BUILD_DIR/scion.iso"
+MINI_ISO="$BUILD_DIR/mini-${FLAVOR}.iso"
+SCION_ISO="$BUILD_DIR/scion-${FLAVOR}.iso"
 EXTRACT_DIR="$BUILD_DIR/extract"
 INITRD_EXTRAS="$BUILD_DIR/initrd.d"
 
 log() {
-    echo "========> $@"
+    echo "$(tput setaf 15) ========> $@ $(tput sgr0)"
 }
 
 prereq() {
@@ -43,17 +44,22 @@ if [ -n "$pkgs" ]; then
     sudo apt-get install $pkgs || exit 1
 fi
 
+if [ ! -r "${FLAVOR}.cfg" ]; then
+    log "Couldn't find ${FLAVOR}.cfg"
+    exit 1
+fi
+
 mkdir -p "$BUILD_DIR"
 if [ ! -e "$MINI_ISO" ]; then
-    log "Fetching mini.iso"
-    wget -nv http://archive.ubuntu.com/ubuntu/dists/trusty/main/installer-amd64/current/images/netboot/mini.iso -O "$MINI_ISO"
+    log "Fetching $MINI_ISO"
+    wget -nv http://archive.ubuntu.com/ubuntu/dists/${FLAVOR}/main/installer-amd64/current/images/netboot/mini.iso -O "$MINI_ISO"
 fi
 
 mkdir -p "$BUILD_DIR"
 [ -e "$EXTRACT_DIR" ] && rm -r "$EXTRACT_DIR"
 [ -e "$INITRD_EXTRAS" ] && rm -r "$INITRD_EXTRAS"
 
-log "Extracting mini.iso"
+log "Extracting $MINI_ISO"
 7z x -o"$EXTRACT_DIR" "$MINI_ISO" > /dev/null || exit 1
 
 log "Extracting initrd.gz"
@@ -61,7 +67,7 @@ gunzip -c "$EXTRACT_DIR/initrd.gz" > "$BUILD_DIR/initrd"
 
 log "Patching initrd.gz"
 mkdir -p "$INITRD_EXTRAS/scion"
-cp trusty.cfg "$INITRD_EXTRAS/preseed.cfg"
+cp ${FLAVOR}.cfg "$INITRD_EXTRAS/preseed.cfg"
 cp common.sh early_command late_command "$INITRD_EXTRAS/scion"
 gen_passwd root.passwd
 gen_passwd scion.passwd
@@ -74,14 +80,14 @@ cp isolinux.cfg "$EXTRACT_DIR/"
 cp txt.cfg "$EXTRACT_DIR/"
 rm -r "$EXTRACT_DIR/[BOOT]"
 
-log "Creating custom ISO image"
+log "Creating ${SCION_ISO}"
 genisoimage -r -J -l \
-    -V "SCION Ubuntu 14.04 Install CD" \
+    -V "SCION Ubuntu ${FLAVOR} Install CD" \
     -b isolinux.bin -c boot.cat -no-emul-boot \
     -boot-load-size 4 -boot-info-table \
     -input-charset default \
-    -o "$SCION_ISO" \
+    -o "${SCION_ISO}" \
     "$EXTRACT_DIR"
 
-log "Post-processing custom image with isohybrid"
-isohybrid "$SCION_ISO"
+log "Post-processing ${SCION_ISO} with isohybrid"
+isohybrid "${SCION_ISO}"
